@@ -5,9 +5,9 @@ import com.makers.princemaker.type.PrinceLevel;
 import com.makers.princemaker.type.SkillType;
 import lombok.*;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 
 /**
  * @author Snow
